Fixed the physics a bit.
git-svn-id: http://www.nlogn.org/svn/epsilon/trunk@41 189e0cc2-7543-4f94-9523-152a149c9a55

+++ Epsilon - test/src/epsilon/game/Physics.java
@@ -14,7 +14,7 @@ public class Physics {
     // environment variables
     public static double GRAVITY = -490.0;
 
-    public static double calculateGravity(int posY, int pposY, int delta) {
+    public static double calculateGravity(double posY, double pposY, int delta) {
         double dt = delta * 0.001;
         double velocity = (posY-pposY)/(dt);
         double dif = -velocity * dt + GRAVITY * dt * dt;
